Removed menu-config.hbs. It's pretty, but it's adding unwanted complication.
Moved main.mjs to root level. Updated all dependent pieces.

Began breaking apart and organizing menu configs and settings initialization.
