+++ jsconfig.json
@@ -9,11 +9,12 @@
   },
   "exclude": ["node_modules", "**/node_modules/*"],
   "include": [
-    "scripts/main.js", 
+    "main.mjs",
     "foundry/dnd5e/**/*.mjs",
-    "foundry/client/client.mjs", 
-    "foundry/client/global.d.mts", 
-    "foundry/common/primitives/global.d.mts"],
+    "foundry/client/client.mjs",
+    "foundry/client/global.d.mts",
+    "foundry/common/primitives/global.d.mts"
+  ],
   "typeAcquisition": {
     "include": ["jquery"]
   }
