--- templates/menu-config.hbs
@@ -1,23 +1,0 @@
-<div class="reference-form">
-  <div class="menu-header">
-    <img src="icons/svg/d20-highlight.svg" alt="Icon" width="96" height="96" />
-    <div class="title-description">
-      <h2>{{localize "DND.MENU.TITLE"}}</h2>
-      <p>{{localize "DND.MENU.DESCRIPTION"}}</p>
-    </div>
-  </div>
-  <hr class="ampersand">
-  <div class="menu-config">
-    {{#each categories}}
-    <div class="form-group">
-      <div class="menu-item">
-        <label class="menu-label" for="{{this.id}}">
-          <span class="menu-name">{{this.name}}</span>
-          <input type="checkbox" name="show{{this.id}}" id="{{this.id}}" {{#if this.checked}}checked{{/if}}>
-          <span class="toggle-indicator"></span>
-        </label>
-      </div>
-    </div>
-    {{/each}}
-  </div>
-</div>
